add bayesian_network
