+++ bayesian_network/src/bayesian_network/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from . import discrete
