+++ bayesian_network/src/bayesian_network/util.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Author: Yuki Furuta <furushchev@jsk.imi.i.u-tokyo.ac.jp>
+
+
+class RObject(object):
+    def __init__(self, debug=False):
+        self.r = R(use_numpy=True)
+        self.debug = debug
+    def rcmd(self, cmd):
+        if self.debug:
+            print(cmd)
+        res = self.r(cmd)
+        if 'Error' in res:
+            raise Exception(res)
+        return res
